updated couchbase definition to work with the latest node.js sdk 2.1.2
